requirements.txt
python-telegram-bot==20.3
websocket-client==1.6.0
numpy==1.25.2